+++ src/views/pages/QualificationPeriod.jsx
@@ -18,6 +18,7 @@ import {
     toastErrorMessageWithParameterConfig
 } from "../../config/toastConfig";
 import {connect} from "react-redux";
+import {verifyTokenExpire} from "../../middleware/verifyToken";
 
 const QualificationPeriod = props => {
     document.title = "Satis client - Paramètre délai de qualification";
@@ -46,7 +47,9 @@ const QualificationPeriod = props => {
                 })
             ;
         }
-        fetchData();
+        if (verifyTokenExpire()) {
+            fetchData();
+        }
     }, [appConfig.apiDomaine, NUMBER_ELEMENT_PER_PAGE]);
 
     const filterShowListBySearchValue = (value) => {
@@ -120,36 +123,38 @@ const QualificationPeriod = props => {
     const deletePeriod = (qualificationPeriodId, index) => {
         DeleteConfirmation.fire(confirmDeleteConfig)
             .then((result) => {
-                if (result.value) {
-                    axios.delete(`${appConfig.apiDomaine}/delai-qualification-parameters/${qualificationPeriodId}`)
-                        .then(response => {
-                            const newUnitTypes = [...qualificationPeriods];
-                            newUnitTypes.splice(index, 1);
-                            setQualificationPeriods(newUnitTypes);
-                            if (showList.length > 1) {
-                                setShowList(
-                                    newUnitTypes.slice(
-                                        getEndByPosition(activeNumberPage) - numberPerPage,
-                                        getEndByPosition(activeNumberPage)
-                                    )
-                                );
-                            } else {
-                                setShowList(
-                                    newUnitTypes.slice(
-                                        getEndByPosition(activeNumberPage - 1) - numberPerPage,
-                                        getEndByPosition(activeNumberPage - 1)
-                                    )
-                                );
-                            }
-                            ToastBottomEnd.fire(toastDeleteSuccessMessageConfig);
-                        })
-                        .catch(error => {
-                            if (error.response.data.error)
-                                ToastBottomEnd.fire(toastErrorMessageWithParameterConfig(error.response.data.error));
-                            else
-                                ToastBottomEnd.fire(toastDeleteErrorMessageConfig);
-                        })
-                    ;
+                if (verifyTokenExpire()) {
+                    if (result.value) {
+                        axios.delete(`${appConfig.apiDomaine}/delai-qualification-parameters/${qualificationPeriodId}`)
+                            .then(response => {
+                                const newUnitTypes = [...qualificationPeriods];
+                                newUnitTypes.splice(index, 1);
+                                setQualificationPeriods(newUnitTypes);
+                                if (showList.length > 1) {
+                                    setShowList(
+                                        newUnitTypes.slice(
+                                            getEndByPosition(activeNumberPage) - numberPerPage,
+                                            getEndByPosition(activeNumberPage)
+                                        )
+                                    );
+                                } else {
+                                    setShowList(
+                                        newUnitTypes.slice(
+                                            getEndByPosition(activeNumberPage - 1) - numberPerPage,
+                                            getEndByPosition(activeNumberPage - 1)
+                                        )
+                                    );
+                                }
+                                ToastBottomEnd.fire(toastDeleteSuccessMessageConfig);
+                            })
+                            .catch(error => {
+                                if (error.response.data.error)
+                                    ToastBottomEnd.fire(toastErrorMessageWithParameterConfig(error.response.data.error));
+                                else
+                                    ToastBottomEnd.fire(toastDeleteErrorMessageConfig);
+                            })
+                        ;
+                    }
                 }
             })
         ;
