+++ src/views/pages/Staff.jsx
@@ -20,6 +20,7 @@ import {verifyPermission} from "../../helpers/permission";
 import {AUTH_TOKEN} from "../../constants/token";
 import {NUMBER_ELEMENT_PER_PAGE} from "../../constants/dataTable";
 import ExportButton from "../components/ExportButton";
+import {verifyTokenExpire} from "../../middleware/verifyToken";
 
 loadCss("/assets/plugins/custom/datatables/datatables.bundle.css");
 axios.defaults.headers.common['Authorization'] = AUTH_TOKEN;
@@ -82,7 +83,8 @@ const   Staff = (props) => {
                     console.log("Something is wrong");
                 });
         }
-        fetchData();
+        if (verifyTokenExpire())
+            fetchData();
     }, [endPoint.list, NUMBER_ELEMENT_PER_PAGE]);
 
     const separateStringByComa = (arrayString) => {
@@ -170,33 +172,35 @@ const   Staff = (props) => {
     const deleteStaff = (staffId, index) => {
         DeleteConfirmation.fire(confirmDeleteConfig)
             .then((result) => {
-                if (result.value) {
-                    axios.delete(endPoint.destroy(staffId))
-                        .then(response => {
-                            const newStaffs = [...staffs];
-                            newStaffs.splice(index, 1);
-                            setStaffs(newStaffs);
-                            if (showList.length > 1) {
-                                setShowList(
-                                    newStaffs.slice(
-                                        getEndByPosition(activeNumberPage) - numberPerPage,
-                                        getEndByPosition(activeNumberPage)
-                                    )
-                                );
-                            } else {
-                                setShowList(
-                                    newStaffs.slice(
-                                        getEndByPosition(activeNumberPage - 1) - numberPerPage,
-                                        getEndByPosition(activeNumberPage - 1)
-                                    )
-                                );
-                            }
-                            ToastBottomEnd.fire(toastDeleteSuccessMessageConfig);
-                        })
-                        .catch(error => {
-                            ToastBottomEnd.fire(toastDeleteErrorMessageConfig);
-                        })
-                    ;
+                if (verifyTokenExpire()) {
+                    if (result.value) {
+                        axios.delete(endPoint.destroy(staffId))
+                            .then(response => {
+                                const newStaffs = [...staffs];
+                                newStaffs.splice(index, 1);
+                                setStaffs(newStaffs);
+                                if (showList.length > 1) {
+                                    setShowList(
+                                        newStaffs.slice(
+                                            getEndByPosition(activeNumberPage) - numberPerPage,
+                                            getEndByPosition(activeNumberPage)
+                                        )
+                                    );
+                                } else {
+                                    setShowList(
+                                        newStaffs.slice(
+                                            getEndByPosition(activeNumberPage - 1) - numberPerPage,
+                                            getEndByPosition(activeNumberPage - 1)
+                                        )
+                                    );
+                                }
+                                ToastBottomEnd.fire(toastDeleteSuccessMessageConfig);
+                            })
+                            .catch(error => {
+                                ToastBottomEnd.fire(toastDeleteErrorMessageConfig);
+                            })
+                        ;
+                    }
                 }
             })
         ;
